tpm2: Set cast of calculation to UINT32
Upstream uses a cast to 'int', which is not right, since the datatype of
availRAMSpaxe is a UINT32. Therefore, set the cast of the calculation to
UINT32.

+++ src/tpm2/NvDynamic.c
@@ -1856,7 +1856,7 @@ NvCapGetCounterAvail(void)
             availNVSpace -= reserved;
     }
     // Compute the available space in RAM
-    availRAMSpace = (RAM_ORDERLY_END - NvRamGetEnd());	/* kgold - removed cast */
+    availRAMSpace = (UINT32)(RAM_ORDERLY_END - NvRamGetEnd());	/* libtpms: fixed cast from 'int' */
 
     // Return the min of counter number in NV and in RAM
     if(availNVSpace / NV_INDEX_COUNTER_SIZE
